Added SERIAL_DELAY Constant
This constant should be used to control the delay function in the DEBUG loop.  If servos are attached while in debug mode and the delay is too large, the servos will appear to not function properly.  While no servos are attached, a larger delay of near 1 second makes the serial monitor output more readable.